Move utils to helper class
