+++ helper/saver.py
@@ -4,7 +4,7 @@ import torch
 from helper import io_utils as io
 from config import FLAGS
 from model.model_factory import create_model
-from utils import create_dir_if_not_exists, get_ts, save, sorted_nicely
+from helper.utils import create_dir_if_not_exists, get_ts, save, sorted_nicely
 
 
 class Saver(object):
