+++ model/model_factory.py
@@ -2,7 +2,7 @@ from collections import Counter
 import torch
 from config import FLAGS
 from model.model_text_gnn import TextGNN
-from utils import parse_as_int_list
+from helper.utils import parse_as_int_list
 
 
 def create_model(dataset):
